Yet another version of CpCallManager::getNewCallId. This one removes the need for using long long int's (and the attendant portability hassles), shortens the Call-IDs slightly, but maintains the other desirable properties of the previous version. Add a test to check that successive Call-IDs aren't too similar.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7854 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXcallLib/include/cp/CpCallManager.h
@@ -881,7 +881,7 @@ private:
     UtlDList mCallList;
     int mLastMetaEventId;
     // Every CallManager shares the same call counter for generating Call-IDs.
-    static INT64 mCallNum;
+    static unsigned int mCallNum;
 
 #endif // DOXYGEN_SHOULD_SKIP_THIS
